+++ src/types.d.ts
@@ -104,16 +104,6 @@ export interface UserCommand {
 }
 
 /**
- * * Object container for parameters of leoIntegration's "apply-selected-node-to-body" method
- */
-export interface ShowBodyParam {
-    node: LeoNode,
-    aside: boolean,
-    showBodyKeepFocus: boolean,
-    force_open?: boolean
-}
-
-/**
  * * Stackable leoBridge actions to be performed by Leo
  */
 export interface LeoAction {
@@ -188,6 +178,8 @@ export interface LeoBridgePackage {
     found?: boolean // find_next, find_previous
     index?: number; // get_all_open_commanders
     language?: string; // get_body_states
+    wrap?: boolean; // get_body_states
+    tabWidth?: number | boolean; // get_body_states either the tabwidth or falsy
     node?: ArchivedPosition; // get_parent, set_opened_file, open_file(s), ?close_file
     children?: ArchivedPosition[]; // get_children
     searchSettings?: LeoGuiFindTabManagerSettings // get_search_settings